Add example for running a remote (stateless) MCP server on Modal (#1265)
It demonstrates how to deploy a simple [MCP server](https://modelcontextprotocol.io/) on Modal, that provides a tool to get the current date and time in a given timezone. It is a stateless MCP server, meaning that it does not store any state between requests, and uses the "streamable HTTP" transport type.

It uses the [FastMCP library](https://github.com/jlowin/fastmcp) to create the MCP server, which is wrapped using FastAPI, and served with a [Modal web endpoint](https://modal.com/docs/guide/webhooks).